feat: BREAKING remove leveldown and memdown
