+++ package.json
@@ -49,8 +49,7 @@
     "ilp-packet": "^3.0.7",
     "ilp-protocol-ccp": "^1.1.1",
     "ilp-protocol-ildcp": "^2.0.0",
-    "leveldown": "^3.0.0",
-    "levelup": "^2.0.1",
+    "levelup": "^2.0.2",
     "lodash": "^4.17.4",
     "long": "^4.0.0",
     "node-fetch": "^2.0.0",
@@ -67,7 +66,7 @@
     "@commitlint/config-conventional": "^6.1.0",
     "@justmoon/json-schema-to-markdown": "^1.0.0",
     "@types/debug": "0.0.30",
-    "@types/levelup": "^3.1.0",
+    "@types/levelup": "^3.1.1",
     "@types/lodash": "^4.14.91",
     "@types/long": "^3.0.32",
     "@types/node": "^11.10.5",
@@ -89,7 +88,6 @@
     "ilp-plugin-mirror": "^1.0.0",
     "json-schema-to-typescript": "^5.2.2",
     "mdast-util-inject": "^1.1.0",
-    "memdown": "^1.4.1",
     "mocha": "^5.0.0",
     "mock-require": "^3.0.1",
     "mock-socket": "^7.1.0",
