+++ src/schemas/Config.json
@@ -251,9 +251,9 @@
       "additionalProperties": true
     },
     "store": {
-      "description": "Name of the store (can be built-in or a require-able module name). Built-in modules are: leveldown, memdown",
+      "description": "Name of the store (can be built-in or a require-able module name). Built-in modules are: memory",
       "type": "string",
-      "default": "leveldown"
+      "default": "memory"
     },
     "storePath": {
       "description": "Shorthand for `config.storeConfig.path`.",
